algorithms/number-of-islands/number-of-islands_test.go
package leetcode

import (
	"testing"
)

func TestNumberofIslands(t *testing.T) {
	cases := []struct {
		desc string
		grid [][]byte
		want int
	}{
		{
			desc: "#1",
			grid: [][]byte{
				{'1', '1', '1', '1', '0'},
				{'1', '1', '0', '1', '0'},
				{'1', '1', '0', '0', '0'},
				{'0', '0', '0', '0', '0'},
			},
			want: 1,
		},
		{
			desc: "#2",
			grid: [][]byte{
				{'1', '1', '0', '0', '0'},
				{'1', '1', '0', '0', '0'},
				{'0', '0', '1', '0', '0'},
				{'0', '0', '0', '1', '1'},
			},
			want: 3,
		},
		{
			desc: "#3",
			grid: [][]byte{
				{'0', '1', '0', '0', '0'},
				{'1', '1', '0', '0', '0'},
				{'0', '0', '1', '0', '0'},
				{'0', '0', '0', '1', '1'},
			},
			want: 3,
		},
	}

	for _, s := range cases {
		t.Run(s.desc, func(t *testing.T) {
			if got := numIslands(s.grid); got != s.want {
				t.Error()
			}
		})
	}
}
